1) Minor backend and frontend fixes to resolve compiling and django management command errors. 2) Temporarily pared down number of Scribe project data models to make migrations for in order to help with local development performance, prior to going to production.

+++ backend/automagic_rest/management/commands/build_data_models.py
@@ -243,7 +243,7 @@ class Command(BaseCommand):
         """
         files_to_delete = glob(f"{root_path}/models/*.py")
         for f in files_to_delete:
-            if not f.endswith("__.py"):
+            if not f.endswith("__.py") and "models.py" not in f:
                 os.remove(f)
 
     def write_schema_files(self, root_path, context):
